+++ BitmovinYospacePlayer/Sources/BitmovinYospacePlayer/API/Config/YospaceSourceConfig.swift
@@ -11,7 +11,7 @@ public class YospaceSourceConfig {
      Initialize a new YospaceSourceConfiguration object.
 
      - Parameters:
-     - yospaceAssetType: YospaceAssetType that tells the Yospace Ad Management SDK if the source is Linear, VOD, or NonLinearStartOver
+     - yospaceAssetType: YospaceAssetType that tells the Yospace Ad Management SDK if the source is Linear, DVR Live, or VOD
      - retryExcludingYospace: Boolean describing if the player should retry the source URL without Yospace when failures occur
      */
     public init(yospaceAssetType: YospaceAssetType, retryExcludingYospace: Bool = false) {