+++ BitmovinYospacePlayer/Sources/BitmovinYospacePlayer/API/Model/YospaceAssetType.swift
@@ -2,6 +2,8 @@ import UIKit
 
 @frozen
 public enum YospaceAssetType: String {
+    @available(*, deprecated, renamed: "dvrLive", message: "Use dvrLive for DVR live playback.")
     case linear
+    case dvrLive
     case vod
 }